requirements.txt
async-generator==1.10
attrs==22.1.0
autobahn==19.11.2
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
dnspython==2.2.1
enum34==1.1.10
exceptiongroup==1.0.0
fonttools==4.38.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kiteconnect==4.1.0
kiwisolver==1.4.4
matplotlib==3.6.1
mibian==0.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.31
NorenRestApiPy @ file:///apps/softwares/py_packages/StarApi-py/dist/NorenRestApiPy-0.0.16-py2.py3-none-any.whl
numpy==1.23.4
outcome==1.2.0
packaging==21.3
pandas==1.5.1
patsy==0.5.3
Pillow==9.2.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==4.3.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
scipy==1.9.3
selenium==4.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
statsmodels==0.13.2
TA-Lib==0.4.25
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
websocket-client==1.4.1
wsproto==1.2.0
zope.interface==5.5.0